Fix syncing issue with Font selector
