+++ src/assets/js/react/components/FontManager/FontManager.js
@@ -4,6 +4,8 @@ import PropTypes from 'prop-types';
 /* Components */
 import FontManagerHeader from './FontManagerHeader';
 import FontManagerBody from './FontManagerBody';
+import { connect } from 'react-redux';
+import { associatedFontManagerSelectBox } from '../../utilities/FontManager/associatedFontManagerSelectBox';
 
 /**
  * @package			Gravity PDF
@@ -66,6 +68,16 @@ export class FontManager extends Component {
 	 */
 	componentWillUnmount() {
 		document.removeEventListener('focus', this.handleFocus, true);
+
+		const { fontList, selectedFont } = this.props;
+		const tabLocation = window.location.search.substring(
+			window.location.search.lastIndexOf('=') + 1
+		);
+
+		/* When closed, ensure font select box has the latest custom font data */
+		if (tabLocation !== 'tools') {
+			return associatedFontManagerSelectBox(fontList, selectedFont);
+		}
 	}
 
 	/**
@@ -111,4 +123,21 @@ export class FontManager extends Component {
 	}
 }
 
-export default FontManager;
+/**
+ * Map redux state to props
+ *
+ * @param { Object } state
+ *
+ * @return {{
+ *  fontList: Array<Object>,
+ *  selectedFont: string,
+ * }} mappedState
+ *
+ * @since 6.12.5
+ */
+const mapStateToProps = (state) => ({
+	fontList: state.fontManager.fontList,
+	selectedFont: state.fontManager.selectedFont,
+});
+
+export default connect(mapStateToProps, {})(FontManager);
